+++ datalink.sql
@@ -2342,8 +2342,21 @@ $$ select datalink.getlength(dlvalue($1,'FS')) $$ language sql;
 comment on function getlength(file_path) is 
   'BFILE - Returns file size';
 
-create or replace function pg_catalog.substr(datalink, pos integer default 1, len integer default 32767) returns text as 
-$$ select datalink.read_text($1,$2,$3) $$ language sql;
+create or replace function pg_catalog.substr(datalink, pos integer default null, len integer default null) 
+ RETURNS text LANGUAGE plpgsql
+AS $$
+begin
+  pos := coalesce(pos,1);
+  if len is not null then 
+    return substr(datalink.read_text($1), pos, len);
+  else
+    if pos > 1 then
+      return substr(datalink.read_text($1), pos);
+    end if;
+  end if;
+  return datalink.read_text($1);
+end
+$$;
 
 ---------------------------------------------------
 -- directories