+++ CHANGELOG.md
@@ -20,6 +20,7 @@ Version 0.24
 - added tables `datalink.dl_access_web` and `datalink.access_web` table and view 
 - added preliminary `has_web_privilege()` functions
 - remove function `instr()`
+- improve function `substr()` to return proper substrings
 
 Version 0.23
 ------------